Moved getSubscribedEvents to trait
