+++ plugins/schemaorg/organization/src/Extension/Organization.php
@@ -62,23 +62,6 @@ final class Organization extends CMSPlugin implements SubscriberInterface
     protected $pluginName = 'Organization';
 
     /**
-     * Returns an array of events this subscriber will listen to.
-     *
-     * @return  array
-     *
-     * @since   4.0.0
-     */
-    public static function getSubscribedEvents(): array
-    {
-        return [
-            'onSchemaPrepareData'                  => 'onSchemaPrepareData',
-            'onSchemaPrepareForm'                  => 'onSchemaPrepareForm',
-            'onSchemaAfterSave'                    => 'onSchemaAfterSave',
-            'onSchemaBeforeCompileHead'            => 'pushSchema',
-        ];
-    }
-
-    /**
      *  Saves the schema to the database
      *
      *  @param   AbstractEvent $event
