+++ plugins/schemaorg/person/src/Extension/Person.php
@@ -62,23 +62,6 @@ final class Person extends CMSPlugin implements SubscriberInterface
     protected $pluginName = 'Person';
 
     /**
-     * Returns an array of events this subscriber will listen to.
-     *
-     * @return  array
-     *
-     * @since   4.0.0
-     */
-    public static function getSubscribedEvents(): array
-    {
-        return [
-            'onSchemaPrepareData' => 'onSchemaPrepareData',
-            'onSchemaPrepareForm' => 'onSchemaPrepareForm',
-            'onSchemaAfterSave' => 'onSchemaAfterSave',
-            'onSchemaBeforeCompileHead' => 'pushSchema',
-        ];
-    }
-
-    /**
      *  Saves the schema to the database
      *
      *  @param   AbstractEvent $event
